tarsier-snapshots/pyproject.toml
[tool.poetry]
name = "tarsier-snapshots"
version = "0.1.0"
description = ""
authors = ["asim-shrestha <asim.shrestha@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bananalyzer = "^0.8.6"
playwright = "^1.41.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
